build(tsup and packaging for npm): this changes the build system to tsup for packaging the widget
package the widget with tsup. its a builder that wraps esbuild and npm-dts. it should be able to
generate the d.ts files however right now they generate as empty. it does build both cjs and esm
builds. we keep the package.json file as is in the package/ folder.